+++ R/design.R
@@ -71,7 +71,7 @@
 #'     points to be added to individual emulators in the bundle.
 #'
 #' See [alm()], [mice()], and [pei()] for examples on customizing `method`. Defaults to [mice()].
-#' @param eval an R function that calculates the customized evaluating metric of the emulator.The function must satisfy the following basic rules:
+#' @param eval an R function that calculates the customized evaluating metric of the emulator. The function must satisfy the following basic rules:
 #' * the first argument is an emulator object that can be either an instance of
 #'   - the `gp` class (produced by [gp()]);
 #'   - the `dgp` class (produced by [dgp()]);
@@ -164,7 +164,6 @@
 #' # load packages and the Python env
 #' library(lhs)
 #' library(dgpsi)
-#' init_py()
 #'
 #' # construct a 2D non-stationary function that takes a matrix as the input
 #' f <- function(x) {
@@ -219,6 +218,10 @@ design <- function(object, N, x_cand, y_cand, n_cand, limits, int, f, reps, freq
 #' @method design gp
 #' @export
 design.gp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, limits = NULL, int = FALSE, f = NULL, reps = 1, freq = c(1, 1), x_test = NULL, y_test = NULL, reset = FALSE, target = NULL, method = mice, eval = NULL, verb = TRUE, check_point = NULL, cores = 1, ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   if ( !inherits(object,"gp") ) stop("'object' must be an instance of the 'gp' class.", call. = FALSE)
 
   N <- check_N(N)
@@ -501,7 +504,9 @@ design.gp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, lim
     if ( run ){
       for ( i in 1:N ){
         if ( n_cand<=length(idx_x_cand) ){
-          idx_sub_cand <- sample(idx_x_cand, n_cand, replace = FALSE)
+          idx_idx <- suppressPackageStartupMessages(clhs::clhs(as.data.frame(x_cand[idx_x_cand,,drop = FALSE]), size = n_cand, progress = FALSE, simple = TRUE))
+          idx_sub_cand <- idx_x_cand[idx_idx]
+          #idx_sub_cand <- sample(idx_x_cand, n_cand, replace = FALSE)
         } else {
           idx_sub_cand <- idx_x_cand
         }
@@ -664,6 +669,10 @@ design.gp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, lim
 #' @method design dgp
 #' @export
 design.dgp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, limits = NULL, int = FALSE, f = NULL, reps = 1, freq = c(1, 1), x_test = NULL, y_test = NULL, reset = FALSE, target = NULL, method = mice, eval = NULL, verb = TRUE, check_point = NULL, cores = 1, train_N = 100, refit_cores = 1, ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   if ( !inherits(object,"dgp") ) stop("'object' must be an instance of the 'dgp' class.", call. = FALSE)
 
   N <- check_N(N)
@@ -964,7 +973,8 @@ design.dgp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, li
     if ( run ){
       for ( i in 1:N ){
         if ( n_cand<=length(idx_x_cand) ){
-          idx_sub_cand <- sample(idx_x_cand, n_cand, replace = FALSE)
+          idx_idx <- suppressPackageStartupMessages(clhs::clhs(as.data.frame(x_cand[idx_x_cand,,drop = FALSE]), size = n_cand, progress = FALSE, simple = TRUE))
+          idx_sub_cand <- idx_x_cand[idx_idx]
         } else {
           idx_sub_cand <- idx_x_cand
         }
@@ -1139,6 +1149,10 @@ design.dgp <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, li
 #' @method design bundle
 #' @export
 design.bundle <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200, limits = NULL, int = FALSE, f = NULL, reps = 1, freq = c(1, 1), x_test = NULL, y_test = NULL, reset = FALSE, target = NULL, method = mice, eval = NULL, verb = TRUE, check_point = NULL, cores = 1, train_N = 100, refit_cores = 1,  ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   if ( !inherits(object,"bundle") ) stop("'object' must be an instance of the 'bundle' class.", call. = FALSE)
 
   N <- check_N(N)
@@ -1642,7 +1656,9 @@ design.bundle <- function(object, N, x_cand = NULL, y_cand = NULL, n_cand = 200,
     if ( run ){
       for ( i in 1:N ){
         if ( n_cand<=length(idx_x_cand) ){
-          idx_sub_cand <- sample(idx_x_cand, n_cand, replace = FALSE)
+          idx_idx <- suppressPackageStartupMessages(clhs::clhs(as.data.frame(x_cand[idx_x_cand,,drop = FALSE]), size = n_cand, progress = FALSE, simple = TRUE))
+          idx_sub_cand <- idx_x_cand[idx_idx]
+          #idx_sub_cand <- sample(idx_x_cand, n_cand, replace = FALSE)
         } else {
           idx_sub_cand <- idx_x_cand
         }