+++ R/pei.R
@@ -58,7 +58,6 @@
 #' # load packages and the Python env
 #' library(lhs)
 #' library(dgpsi)
-#' init_py()
 #'
 #' # construct a 1D non-stationary function
 #' f <- function(x) {
@@ -103,6 +102,10 @@ pei <- function(object, x_cand, ...){
 #' @method pei gp
 #' @export
 pei.gp <- function(object, x_cand, pseudo_points = NULL, batch_size = 1, ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   #check class
   if ( !inherits(object,"gp") ) stop("'object' must be an instance of the 'gp' class.", call. = FALSE)
   training_input <- object$data$X
@@ -165,6 +168,10 @@ pei.gp <- function(object, x_cand, pseudo_points = NULL, batch_size = 1, ...) {
 #' @method pei dgp
 #' @export
 pei.dgp <- function(object, x_cand, pseudo_points = NULL, batch_size = 1, workers = 1, threading = FALSE, aggregate = NULL, ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   #check class
   if ( !inherits(object,"dgp") ) stop("'object' must be an instance of the 'dgp' class.", call. = FALSE)
   if ( object$constructor_obj$all_layer[[object$constructor_obj$n_layer]][[1]]$type == 'likelihood' ){
@@ -280,6 +287,10 @@ pei.dgp <- function(object, x_cand, pseudo_points = NULL, batch_size = 1, worker
 #' @method pei bundle
 #' @export
 pei.bundle <- function(object, x_cand, pseudo_points = NULL, batch_size = 1, workers = 1, threading = FALSE, aggregate = NULL, ...) {
+  if ( is.null(pkg.env$dgpsi) ) {
+    init_py(verb = F)
+    if (pkg.env$restart) return()
+  }
   #check class
   if ( !inherits(object,"bundle") ) stop("'object' must be an instance of the 'bundle' class.", call. = FALSE)
   #check no of emulators