+++ NEWS.md
@@ -10,6 +10,8 @@
 - A new slot called `specs` is added to the objects returned by `gp()` and `dgp()` that contains the key information of the kernel functions used in the constructions of GP and DGP emulators.
 - Due to a bug in the latest version of an underlying Python package, the emulators saved by `write()` in version `2.1.6` and `2.2.0` may not work properly with `update()` and `design()` when they are loaded back by `read()` in this version. This bug has been addressed in this version so emulators saved in this version would not have the compatibility issue in future version.
 - A new sequential design criterion, called the Variance of Improvement for Global Fit (VIGF), is added to the package with the function `vigf()`.
+- The sampling from an existing candidate set `x_cand` in `design()` is changed from a random sampling to a conditioned Latin Hypercube sampling in `clhs` package.
+- The python environment is now automatically installed or invoked when a function from the package is executed. One does not need to run `init_py()` to activate the required python environment but `init_py()` is still useful to re-install and uninstall the underlying python environment. A `verb` argument is added to `init_py()` to switch on/off the trace information.
 
 # dgpsi 2.2.0